Updated project with db connection
Project update with successful db connection and a few change in UI

+++ src/main/kotlin/controllers/employeeController.kt
@@ -92,7 +92,7 @@ class employeeController {
             println("Employee Not Deleted...")
     }
 
-    fun search(id: Long): employeeModel? {
+    fun search(id: Int): employeeModel? {
         var foundEmployee = employees.findOne(id)
         return foundEmployee
     }
