+++ src/main/kotlin/models/employeeModel.kt
@@ -2,8 +2,8 @@ package models
 
 data class employeeModel(
 
-    var id: Long = 0,
-    var fName: String = "",
+    var id: Int = 0,
+    var fName: String= "",
     var sName: String = "",
     var dateOfB: String = "",
     var email: String = "",
