+++ src/main/kotlin/models/employeeStore.kt
@@ -2,7 +2,7 @@ package models
 
 interface employeeStore {
     fun findAll(): List<employeeModel>
-    fun findOne(id: Long): employeeModel?
+    fun findOne(id: Int): employeeModel?
     fun create(employee: employeeModel)
     fun update(employee: employeeModel)
     fun delete(employee: employeeModel)
